Split SamplingProfiler into View-only Data Class and Create Function
SamplingProfiler had two purposes:
1. Hold post processed sampling data (that should be also stored in CaptureData).
2. Create that post processed data (which requires access to CaptureData).

This required a forward declaration CaptureData.

Within this change, the two functionalities are split into two classes:
1. OrbitClientData/PostProcessedSamplingData holds the post processed data.
2. OrbitClientModel/SamplingDataPostProcessor encapsulates the old SamplingProfiler
   and provides a free function two create the PostProcessedSamplingData.

Bug: http://b/166238019
Test: Compile
