+++ OrbitGl/CallTreeView.h
@@ -5,6 +5,7 @@
 #ifndef ORBIT_GL_CALL_TREE_VIEW_H_
 #define ORBIT_GL_CALL_TREE_VIEW_H_
 
+#include "OrbitClientData/PostProcessedSamplingData.h"
 #include "OrbitClientModel/CaptureData.h"
 #include "Path.h"
 #include "absl/container/node_hash_map.h"
@@ -110,10 +111,12 @@ class CallTreeThread : public CallTreeNode {
 class CallTreeView : public CallTreeNode {
  public:
   [[nodiscard]] static std::unique_ptr<CallTreeView> CreateTopDownViewFromSamplingProfiler(
-      const SamplingProfiler& sampling_profiler, const CaptureData& capture_data);
+      const PostProcessedSamplingData& post_processed_sampling_data,
+      const CaptureData& capture_data);
 
   [[nodiscard]] static std::unique_ptr<CallTreeView> CreateBottomUpViewFromSamplingProfiler(
-      const SamplingProfiler& sampling_profiler, const CaptureData& capture_data);
+      const PostProcessedSamplingData& post_processed_sampling_data,
+      const CaptureData& capture_data);
 
   CallTreeView() : CallTreeNode{nullptr} {}
 };
