+++ OrbitGl/EventTrack.cpp
@@ -6,6 +6,7 @@
 
 #include "App.h"
 #include "GlCanvas.h"
+#include "OrbitClientData/PostProcessedSamplingData.h"
 #include "PickingManager.h"
 
 using orbit_client_protos::CallstackEvent;
@@ -88,7 +89,7 @@ void EventTrack::UpdatePrimitives(uint64_t min_tick, uint64_t max_tick, PickingM
       }
     };
 
-    if (thread_id_ == SamplingProfiler::kAllThreadsFakeTid) {
+    if (thread_id_ == PostProcessedSamplingData::kAllThreadsFakeTid) {
       capture_data->GetCallstackData()->ForEachCallstackEvent(action_on_callstack_events);
     } else {
       capture_data->GetCallstackData()->ForEachCallstackEventOfTid(thread_id_,
@@ -120,7 +121,7 @@ void EventTrack::UpdatePrimitives(uint64_t min_tick, uint64_t max_tick, PickingM
         batcher->AddShadedBox(pos, size, z, kGreenSelection, std::move(user_data));
       }
     };
-    if (thread_id_ == SamplingProfiler::kAllThreadsFakeTid) {
+    if (thread_id_ == PostProcessedSamplingData::kAllThreadsFakeTid) {
       capture_data->GetCallstackData()->ForEachCallstackEvent(action_on_callstack_events);
     } else {
       capture_data->GetCallstackData()->ForEachCallstackEventOfTid(thread_id_,
@@ -165,7 +166,7 @@ bool EventTrack::IsEmpty() const {
   const CaptureData* capture_data = time_graph_->GetCaptureData();
   if (capture_data == nullptr) return true;
   const uint32_t callstack_count =
-      (thread_id_ == SamplingProfiler::kAllThreadsFakeTid)
+      (thread_id_ == PostProcessedSamplingData::kAllThreadsFakeTid)
           ? capture_data->GetCallstackData()->GetCallstackEventsCount()
           : capture_data->GetCallstackData()->GetCallstackEventsOfTidCount(thread_id_);
   return callstack_count == 0;
