+++ OrbitGl/GpuTrack.cpp
@@ -7,6 +7,7 @@
 #include "App.h"
 #include "GlCanvas.h"
 #include "OrbitBase/Profiling.h"
+#include "OrbitClientData/PostProcessedSamplingData.h"
 #include "TimeGraph.h"
 #include "absl/strings/str_format.h"
 
@@ -54,7 +55,8 @@ bool GpuTrack::IsTimerActive(const TimerInfo& timer_info) const {
   // We do not properly track the PID for GPU jobs and we still want to show
   // all jobs as active when no thread is selected, so this logic is a bit
   // different than SchedulerTrack::IsTimerActive.
-  bool no_thread_selected = GOrbitApp->selected_thread_id() == SamplingProfiler::kAllThreadsFakeTid;
+  bool no_thread_selected =
+      GOrbitApp->selected_thread_id() == PostProcessedSamplingData::kAllThreadsFakeTid;
 
   return is_same_tid_as_selected || no_thread_selected;
 }
