+++ OrbitGl/SamplingReportDataView.cpp
@@ -276,7 +276,7 @@ void SamplingReportDataView::SetSampledFunctions(const std::vector<SampledFuncti
 
 void SamplingReportDataView::SetThreadID(ThreadID tid) {
   tid_ = tid;
-  if (tid == SamplingProfiler::kAllThreadsFakeTid) {
+  if (tid == PostProcessedSamplingData::kAllThreadsFakeTid) {
     name_ = absl::StrFormat("%s\n(all threads)", GOrbitApp->GetCaptureData().process_name());
   } else {
     name_ = absl::StrFormat("%s\n[%d]", GOrbitApp->GetCaptureData().GetThreadName(tid_), tid_);
