+++ OrbitQt/CallTreeViewItemModel.cpp
@@ -4,6 +4,8 @@
 
 #include "CallTreeViewItemModel.h"
 
+#include "OrbitClientData/PostProcessedSamplingData.h"
+
 CallTreeViewItemModel::CallTreeViewItemModel(std::unique_ptr<CallTreeView> call_tree_view,
                                              QObject* parent)
     : QAbstractItemModel{parent}, call_tree_view_{std::move(call_tree_view)} {}
@@ -16,7 +18,7 @@ QVariant CallTreeViewItemModel::GetDisplayRoleData(const QModelIndex& index) con
   if (thread_item != nullptr) {
     switch (index.column()) {
       case kThreadOrFunction:
-        if (thread_item->thread_id() == SamplingProfiler::kAllThreadsFakeTid) {
+        if (thread_item->thread_id() == PostProcessedSamplingData::kAllThreadsFakeTid) {
           return QString::fromStdString(
               thread_item->thread_name().empty()
                   ? "(all threads)"
